Got rid of macros from vector math host vector.
