+++ core/src/cpp/Makefile.x64_sse_linux
@@ -5,7 +5,7 @@ BIN      = bin/interpreter_x64
 
 # Compiler settings
 CXXFLAGS = --std=c++17 -Wall -Werror -Ofast -march=native -mtune=native -mfpmath=sse -msse2 -fPIC -pipe -Iinclude
-GCC_CXXFLAGS = -flto -fexpensive-optimizations -funroll-all-loops -DMESSAGE='"Compiled with GCC"'
+GCC_CXXFLAGS = -fexpensive-optimizations -funroll-all-loops -DMESSAGE='"Compiled with GCC"'
 CLANG_CXXFLAGS = -funroll-loops -DMESSAGE='"Compiled with Clang"'
 UNKNOWN_CXXFLAGS = -DMESSAGE='"Compiled with an unknown compiler"'
