+++ core/src/cpp/include/host/vector/mat2x2.hpp
@@ -14,16 +14,84 @@
  *    - 64-bit 680x0-inspired Virtual Machine and assembler -
  */
 
-#include "templates.hpp"
 #include <machine/register.hpp>
+#include "matrix_common.hpp"
+
+#ifdef MATRIX_FORCE_DOUBLE
+#define MT T
+#else
+#define MT float64
+#endif
 
 namespace MC64K::StandardTestHost::VectorMath {
 
+/**
+ * Mat2x2 multiplication.
+ *
+ * C = A x B, A is by row, B is by column
+ */
+template<typename T>
+inline void mat2x2_multiply(T* pfDst, T const* pfSrcA, T const* pfSrcB) {
+    // Row 1
+    pfDst[M2_11] = (T)(
+        ((MT)pfSrcA[M2_11] * (MT)pfSrcB[M2_11]) +
+        ((MT)pfSrcA[M2_12] * (MT)pfSrcB[M2_21])
+    );
+    pfDst[M2_12] = (T)(
+        ((MT)pfSrcA[M2_11] * (MT)pfSrcB[M2_12]) +
+        ((MT)pfSrcA[M2_12] * (MT)pfSrcB[M2_22])
+    );
+
+    // Row 2
+    pfDst[M2_21] = (T)(
+        ((MT)pfSrcA[M2_21] * (MT)pfSrcB[M2_11]) +
+        ((MT)pfSrcA[M2_22] * (MT)pfSrcB[M2_21])
+    );
+    pfDst[M2_22] = (T)(
+        ((MT)pfSrcA[M2_21] * (MT)pfSrcB[M2_12]) +
+        ((MT)pfSrcA[M2_22] * (MT)pfSrcB[M2_22])
+    );
+}
+
+template<typename T>
+inline void mat2x2_transpose(T* pDst, T const* pSrc) {
+    pDst[M2_11] = pSrc[M2_11];
+    pDst[M2_12] = pSrc[M2_21];
+    pDst[M2_21] = pSrc[M2_12];
+    pDst[M2_22] = pSrc[M2_22];
+}
+
+template<typename T>
+inline T mat2x2_determinant(T const* pfSrc) {
+    return pfSrc[M2_11] * pfSrc[M2_22] - pfSrc[M2_12] * pfSrc[M2_21];
+}
+
+template<typename T>
+inline uint64 mat2x2_inverse(T* pfDst, T const* pfSrc) {
+    T fDeterminant = mat2x2_determinant<T>(pfSrc);
+    if (fDeterminant) {
+        fDeterminant = (T)1.0 / fDeterminant;
+        pfDst[M2_11] = fDeterminant * pfSrc[M2_22];
+        pfDst[M2_22] = fDeterminant * pfSrc[M2_11];
+        pfDst[M2_12] = -fDeterminant * pfSrc[M2_12];
+        pfDst[M2_21] = -fDeterminant * pfSrc[M2_21];
+        return ABI::ERR_NONE;
+    } else {
+        return ERR_ZERO_DIVIDE;
+    }
+}
+
+/**
+ * m2x2(a0) = I
+ */
 template<typename T>
 inline void m2x2_identity() {
     mat_identity<T, 2>(Interpreter::gpr<ABI::PTR_REG_0>().address<T>());
 }
 
+/**
+ * m2x2(a1) = m2x2(a0)
+ */
 template<typename T>
 inline void m2x2_copy() {
     mat_copy<T, 2>(
@@ -32,6 +100,9 @@ inline void m2x2_copy() {
     );
 }
 
+/**
+ * m2x2(a0) *= fp0
+ */
 template<typename T>
 inline void m2x2_scale_assign() {
     mat_scale_assign<T, 2>(
@@ -40,6 +111,9 @@ inline void m2x2_scale_assign() {
     );
 }
 
+/**
+ * m2x2(a1) = fp0 * m2x2(a0)
+ */
 template<typename T>
 inline void m2x2_scale() {
     mat_scale<T, 2>(
@@ -49,6 +123,9 @@ inline void m2x2_scale() {
     );
 }
 
+/**
+ * m2x2(a1) += m2x2(a0)
+ */
 template<typename T>
 inline void m2x2_add_assign() {
     mat_add_assign<T, 2>(
@@ -57,6 +134,9 @@ inline void m2x2_add_assign() {
     );
 }
 
+/**
+ * m2x2(a2) = m2x2(a1) + m2x2(a0)
+ */
 template<typename T>
 inline void m2x2_add() {
     mat_add<T, 2>(
@@ -66,6 +146,9 @@ inline void m2x2_add() {
     );
 }
 
+/**
+ * m2x2(a1) -= m2x2(a0)
+ */
 template<typename T>
 inline void m2x2_sub_assign() {
     mat_sub_assign<T, 2>(
@@ -74,6 +157,9 @@ inline void m2x2_sub_assign() {
     );
 }
 
+/**
+ * m2x2(a2) = m2x2(a1) - m2x2(a0)
+ */
 template<typename T>
 inline void m2x2_sub() {
     mat_sub<T, 2>(
@@ -83,6 +169,9 @@ inline void m2x2_sub() {
     );
 }
 
+/**
+ * m2x2(a2) = m2x2(a1) x m2x2(a0)
+ */
 template<typename T>
 inline void m2x2_multiply() {
     mat2x2_multiply<T>(
@@ -92,6 +181,9 @@ inline void m2x2_multiply() {
     );
 }
 
+/**
+ * m2x2(a1) x= m2x2(a0)
+ */
 template<typename T>
 inline void m2x2_multiply_assign() {
     T* pfDst = Interpreter::gpr<ABI::PTR_REG_1>().address<T>();
@@ -104,6 +196,9 @@ inline void m2x2_multiply_assign() {
     );
 }
 
+/**
+ * m2x2(a1) = m2x2^T(a0)
+ */
 template<typename T>
 inline void m2x2_transpose() {
     mat2x2_transpose<T>(
@@ -112,6 +207,9 @@ inline void m2x2_transpose() {
     );
 }
 
+/**
+ * fp0 = det m2x2(a0)
+ */
 template<typename T>
 inline void m2x2_determinant() {
     Interpreter::fpr<ABI::FLT_REG_0>().value<T>() = mat2x2_determinant<T>(
@@ -119,6 +217,9 @@ inline void m2x2_determinant() {
     );
 }
 
+/**
+ * m2x2(a1) = m2x2^-1(a0)
+ */
 template<typename T>
 inline void m2x2_inverse() {
     Interpreter::gpr<ABI::INT_REG_0>().uQuad = mat2x2_inverse<T>(
@@ -127,19 +228,8 @@ inline void m2x2_inverse() {
     );
 }
 
-/**
- * Applies a Mat2x2 to an input set of Vec2
- */
-template<typename T>
-inline void vec2_transform_2x2(T* pfDst, T const* pfSrc, T const* pfM, size_t uCount) {
-    while (uCount--) {
-        *pfDst++ = pfM[M2_11] * pfSrc[V_X] + pfM[M2_12] * pfSrc[V_Y];
-        *pfDst++ = pfM[M2_21] * pfSrc[V_X] + pfM[M2_22] * pfSrc[V_Y];
-        pfSrc += 2;
-    }
 }
 
-}
-
+#undef MT
 #endif
